improved prompts + modularity
